rocky-agent/requirements.txt
faster-whisper>=1.0.0
piper-tts>=1.2.0
sounddevice>=0.4.6
numpy>=1.24
requests>=2.28
pynput>=1.7.6
pyyaml>=6.0
ollama>=0.4.0
